refactor: deduplicate canMergeRects and mergeRects functions (#1282)
Extract duplicate functions to shared utils module:
- Created src/core/dirtyRects/utils.ts with canMergeRects and mergeRects
- Updated src/core/dirtyRects.ts to import from utils
- Updated src/core/dirtyRects/regions.ts to import from utils
- Removed duplicate function definitions

This resolves the exact duplicates found in the Semfora audit between
the legacy monolithic dirtyRects.ts and the modular dirtyRects/
structure. No circular dependencies introduced - utils.ts only imports
types from types.ts.
